Convert "Projects" components to typescript
