+++ src/components/projects/ProjectDonationCard/styles.ts
@@ -1,9 +1,6 @@
-/* eslint-disable react/display-name */
-import React from 'react';
 import styled from 'styled-components';
-import { FiCreditCard, FiXCircle } from 'react-icons/fi';
 
-const DonationCard = styled.div`
+export const DonationCard = styled.div`
   position: relative;
   background: ${(props) => props.theme.palette.paper};
   width: 100%;
@@ -58,27 +55,3 @@ const DonationCard = styled.div`
     color: ${(props) => props.theme.palette.text.secondary};
   }
 `;
-
-const ProjectDonationCard = ({ donation, showDeleteButton, onDelete }) => {
-  const getIcon = () => {
-    return <FiCreditCard />;
-  };
-
-  return (
-    <DonationCard>
-      {showDeleteButton && (
-        <button type="button" onClick={onDelete}>
-          <FiXCircle />
-        </button>
-      )}
-      <div className="icon">{getIcon(donation.type)}</div>
-
-      <div>
-        <h4>{donation.title}</h4>
-        <p>{donation.description}</p>
-      </div>
-    </DonationCard>
-  );
-};
-
-export default ProjectDonationCard;
